yet an other failing test: test the alpha function
svn path=/trunk/koffice/; revision=1010792

+++ krita/plugins/colorspaces/ctlcs/tests/KoRgb32fTest.cpp
@@ -45,9 +45,9 @@ void KoRgb32fTest::testConversion()
   
   // Test alpha function
   p32f.alpha = 1.0;
-  QCOMPARE(rgb32f->alpha(p32fPtr), 255);
+  QCOMPARE( qint32(rgb32f->alpha(p32fPtr)), 255);
   p32f.alpha = 0.5;
-  QCOMPARE(rgb32f->alpha(p32fPtr), 125);
+  QCOMPARE( qint32(rgb32f->alpha(p32fPtr)), 125);
   
   // Test conversion of black from 32f to 16u back to 32f
   p32f.red = 0.0;
